Widget refactoring
Cleanup the namespaces for widgets and reduce the amount
of code placed in `gui.py`
