+++ cloudinstall/ui/widgets/infodialog.py
@@ -13,11 +13,20 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+from urwid import (WidgetWrap, Button, LineBox, Pile, Text)
 
-from .unit import UnitInfoWidget  # NOQA
-from .statusbar import StatusBarWidget  # noqa
-from .infodialog import InfoDialogWidget  # noqa
-from .selectordialog import SelectorWithDescriptionWidget  # noqa
-from .passwordinput import PasswordInput  # noqa
-from .maasinput import MaasServerInput  # noqa
-from .landscapeinput import LandscapeInput  # noqa
+
+class InfoDialogWidget(WidgetWrap):
+
+    """A widget that displays a message and a close button."""
+
+    def __init__(self, message, close_func):
+        self.close_func = close_func
+        button = Button("Close", self.do_close)
+        box = LineBox(Pile([Text(message),
+                            button]),
+                      title="Info")
+        super().__init__(box)
+
+    def do_close(self, sender):
+        self.close_func(self)
